+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to the Backend and SPA Admin will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [1.5.1] - 2025-12-09
 
 ### Added
 - Added bed config to the manual booking stepper.
+- Added editable bed configuration dropdown in the SPA Admin.
+- Added “Update” action to save the bed configuration.
+- Added backend endpoint to update the bed configuration.
+- Added form validation, error handling and try/catch in the update controller.
 
 
 ## [1.5.0] - 2025-12-08